admin: move admin client to admin/client
